Wrote the start of RSpec section; completely refactored the ComplexNumber class and wrote more unit tests about it; implemented methods square root (such difficult, much mind boggling, wow), reciprocal, and conjugate.
