+++ code/unit_tests/complex/test/test_handmade_simple.rb
@@ -1,9 +1,10 @@
 #!/usr/bin/env ruby
 
-require_relative "complex.rb"
+require_relative "../lib/complex.rb"
+#~ require_relative "../lib/operations.rb"
 
-case1 = ComplexNumber.new("-17", "-4i")
-case2 = ComplexNumber.new("0", "2i")
+case1 = Rex::ComplexNumber.new("-17", "-4i")
+case2 = Rex::ComplexNumber.new("0", "2i")
 
 puts "ComplexNumber class test suite"
 puts
@@ -15,5 +16,8 @@ puts "(#{case1}) + (#{case2}) = #{case1 + case2}"
 puts "(#{case1}) - (#{case2}) = #{case1 - case2}"
 puts "(#{case1}) * (#{case2}) = #{case1 * case2}"
 puts "(#{case1}) / (#{case2}) = #{case1 / case2}"
+puts "sqrt(#{case1}) = #{case1.sqrt}"
+puts "conjugate(#{case1}) = #{case1.conjugate}"
+puts "reciprocal(#{case1}) = #{case1.reciprocal}"
 puts
 puts "Done."
